feat: initial button module functionality
Need to think more about how to structure functionality specific to a chat frame versus
functionality for the entire chat frame. Initial button module will allow players to show or hide
the navigational buttons as they are specific to a chat frame. That is how the addon functions
currently, so it's easy to implement this slice.