fix(db): invalid database index error dependent on New Whispers option
If a user sets the option New Whispers to Both or New Tab and receives or sends a whisper, a new
chat frame is created. No table in the database exists for the new chat frame because a table is
only created for the chat frames that exist when the addon first loads. This fix creates a table in
the database for new chat frames when GetChatFramesTable gets called.

issue #15